requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-telegram-bot==21.0.1
asyncpg==0.29.0
pgvector==0.2.5
sqlalchemy[asyncio]==2.0.25
python-dotenv==1.0.1
google-generativeai==0.4.1
httpx==0.26.0
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
numpy==1.26.4